+++ sway/input/seatop_move_floating.c
@@ -8,45 +8,44 @@ struct seatop_move_floating_event {
 	struct sway_container *con;
 };
 
-static void handle_motion(struct sway_seat *seat, uint32_t time_msec) {
-	struct seatop_move_floating_event *e = seat->seatop_data;
-	desktop_damage_whole_container(e->con);
-	container_floating_translate(e->con,
-			seat->cursor->cursor->x - seat->cursor->previous.x,
-			seat->cursor->cursor->y - seat->cursor->previous.y);
-	desktop_damage_whole_container(e->con);
+static void handle_button(struct sway_seat *seat, uint32_t time_msec,
+		struct wlr_input_device *device, uint32_t button,
+		enum wlr_button_state state) {
+	if (seat->cursor->pressed_button_count == 0) {
+		struct seatop_move_floating_event *e = seat->seatop_data;
+
+		// We "move" the container to its own location
+		// so it discovers its output again.
+		container_floating_move_to(e->con, e->con->x, e->con->y);
+
+		seatop_begin_default(seat);
+	}
 }
 
-static void handle_finish(struct sway_seat *seat, uint32_t time_msec) {
+static void handle_motion(struct sway_seat *seat, uint32_t time_msec,
+		double dx, double dy) {
 	struct seatop_move_floating_event *e = seat->seatop_data;
-
-	// We "move" the container to its own location
-	// so it discovers its output again.
-	container_floating_move_to(e->con, e->con->x, e->con->y);
-	cursor_set_image(seat->cursor, "left_ptr", NULL);
-}
-
-static void handle_abort(struct sway_seat *seat) {
-	cursor_set_image(seat->cursor, "left_ptr", NULL);
+	desktop_damage_whole_container(e->con);
+	container_floating_translate(e->con, dx, dy);
+	desktop_damage_whole_container(e->con);
 }
 
 static void handle_unref(struct sway_seat *seat, struct sway_container *con) {
 	struct seatop_move_floating_event *e = seat->seatop_data;
 	if (e->con == con) {
-		seatop_abort(seat);
+		seatop_begin_default(seat);
 	}
 }
 
 static const struct sway_seatop_impl seatop_impl = {
+	.button = handle_button,
 	.motion = handle_motion,
-	.finish = handle_finish,
-	.abort = handle_abort,
 	.unref = handle_unref,
 };
 
 void seatop_begin_move_floating(struct sway_seat *seat,
-		struct sway_container *con, uint32_t button) {
-	seatop_abort(seat);
+		struct sway_container *con) {
+	seatop_end(seat);
 
 	struct seatop_move_floating_event *e =
 		calloc(1, sizeof(struct seatop_move_floating_event));
@@ -57,7 +56,6 @@ void seatop_begin_move_floating(struct sway_seat *seat,
 
 	seat->seatop_impl = &seatop_impl;
 	seat->seatop_data = e;
-	seat->seatop_button = button;
 
 	container_raise_floating(con);
